refactor: move api directory to lib
