+++ components/challenge/ChallengeBrowseClient.tsx
@@ -7,8 +7,8 @@ import toast from 'react-hot-toast';
 import ChallengeCard from "./ChallengeCard";
 import { Challenge, ChallengeCreateRequest, ChallengeFormData } from "@/types/Challenge";
 import { CategoryDto } from "@/types/Category";
-import { fetchAllChallenges, fetchChallengesByCategory, createChallenge } from "@/api/challenge";
-import { fetchCategories } from "@/api/category";
+import { fetchAllChallenges, fetchChallengesByCategory, createChallenge } from "@/lib/api/challenge";
+import { fetchCategories } from "@/lib/api/category";
 import Sidebar from "../common/Sidebar";
 import BottomNav from "../common/BottomNav";
 import ChallengeApplyModal from "./ChallengeApplyModal";
