+++ lib/api/auth.ts
