+++ lib/api/signup.ts
